ai level 3

+++ src/moves/king_move.h
@@ -8,7 +8,7 @@ class KingMove final : public AbstractMove {
 	Board *board;
 	std::string from, to;
 
-	size_t capture_count;
+	int capture_count;
 	std::shared_ptr<Piece> captured;
 
 	std::string rook_loc, rook_dest;
